requirements.txt
streamlit==1.45.0
alpaca-py==0.40.0
python-dotenv==1.1.0
